remove redundant presser mod
